Modifications for hard patch selection and question embedding input
Modified LSTM unrolling loop to calculate a patch selection index
which is passed to the LSTM at the subsequent timestep to select a
patch. Modified input to be concatenation of question embedding,
lo-rez image, and convolution of selected patch. Modified output
of final timestep to be LSTM hidden state (for passing on to
classifier). Modified model building code in LSTM_DNN.py to handle
changes to networks.py. Modified combination network to no longer
accept question embedding input; however, I did not otherwise change
names or method signatures to make more sense in light of the lack
of any combination now.